🗃️  [DB]: Prisma 스키마, 서비스 수정 및 모듈 생성
