--- src/prisma/prisma.service.ts
@@ -1,14 +1,0 @@
-import { Injectable, OnModuleInit, OnModuleDestroy } from '@nestjs/common';
-import { PrismaClient } from '@prisma/client';
-
-
-@Injectable()
-export class PrismaService extends PrismaClient implements OnModuleInit {
-  async onModuleInit() {
-  await this.$connect();
-    }
-
-async onModuleDestroy() {
-await this.$disconnect();
-    }
-}
